делает ДЗ 5
